backend/requirements.txt
fastapi
httpx
uvicorn
psutil
numpy
pydantic
python-docx
PyPDF2
aiofiles
Pillow
websockets
